// src/os/dto/create.os.dto.ts
export interface OS_DTO {
  idOS: string;
  cliente: string;
  operador: string;
  tipoOS: string;
  zona: string;
  equipe: string;
  transporte: string;
  dataEdicao: string;
  dataInsercao: string;
  dataAbertura: string;
  dataFechamento: string;
  taxa: string;
  correcao: string;
  trs: string;
  trsMin: number;
  pontos: number;
}